Working on database
All jsp classes should be insertted (minus a few minor ones that may need to be added later)
Started working on database classes. Finished majority of the car database and working on Member database. Working off of the software pc database (connect to it using teamviewer if you need to test, or create a local database following same names)
Need to create a hashing algorithm for this and test all the database classes